requirements.txt
python-telegram-bot==20.3
pytesseract
pdf2image
python-docx
arabic-reshaper
python-bidi
pdfplumber
requests
Pillow
